+++ src/migrators/LinkitMigrator.php
@@ -43,7 +43,7 @@ class LinkitMigrator extends BaseMigrator
         return self::TYPE_MAP[$sourceType] ?? null;
     }
 
-    protected function migrateField(object $field): bool
+    protected function migrateField(array $field): bool
     {
         $this->log("Migrating Linkit field: {$field['handle']} (ID: {$field['id']})");
 
@@ -56,6 +56,10 @@ class LinkitMigrator extends BaseMigrator
         return true;
     }
 
+    /**
+     * @param array<string, mixed> $linkitSettings
+     * @return array<string, mixed>
+     */
     private function convertSettings(array $linkitSettings): array
     {
         $linkTypes = [];